add the indexer grpc to fullnode.
